Get rid of std_stable_sorter.
The class std_stable_sorter was a workaround to bypass the fact that there
was no way to properly handle sorters in the library that had both a
stable and an unstable version. This commit removes std_stable_sorter from
the library and introduces stable_adapter<std_sorter> to replace it.
