+++ include/cpp-sort/sorters.h
@@ -41,7 +41,6 @@
 #include <cpp-sort/sorters/smooth_sorter.h>
 #include <cpp-sort/sorters/spread_sorter.h>
 #include <cpp-sort/sorters/std_sorter.h>
-#include <cpp-sort/sorters/std_stable_sorter.h>
 #include <cpp-sort/sorters/tim_sorter.h>
 #include <cpp-sort/sorters/verge_sorter.h>
